+++ ARTX-PoliceScience/ARTX-PoliceScience/ViewController/QuizViewController.swift
@@ -9,10 +9,28 @@ import UIKit
 
 class QuizViewController: UIViewController {
     
-    private let titleView = TitleView()
-    private let progressbarView = ProgressbarView()
+    let partNumber: Int
+    let partTitle: String
+    
+    private var currentQuizNumber: Int = 0
+    
+    private let titleView = QuizTitleView()
+    private let progressbarView = QuizProgressView()
     private let quizView = QuizView()
     private let oxbuttonView = OXbuttonView()
+    private let viewmodel: QuizViewModel
+    
+    init(partNumber: Int, partTitle: String, chapter: Chapter, currentQuizNumber: Int) {
+        self.partNumber = partNumber
+        self.partTitle = partTitle
+        self.viewmodel = QuizViewModel(chapter: chapter)
+        self.currentQuizNumber = currentQuizNumber
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -23,6 +41,7 @@ class QuizViewController: UIViewController {
         
         
         layout()
+        update()
     }
     
     private func layout() {
@@ -95,10 +114,18 @@ class QuizViewController: UIViewController {
         ])
     }
     
+    private func update() {
+        quizView.quizLabel.text = viewmodel.question(to: self.currentQuizNumber)
+    }
+    
 }
 
 extension QuizViewController {
     @objc func backButtonTapped() {
         
+    }
+    
+    @objc func nextQuizButtonTapped() {
+        self.currentQuizNumber += 1
     }
 }
